All Parallel Append commits until now together: -> BUG FIX: Check NULL before dereference the point -> WIP: Force APPEND always require motion from its children -> WIP: Add CDistributionSpecStrictHashed, compiled
Failed hash redistribution test

We suspect the issue is on the matching of new strict hash distribution
with the CPhysicalMotionHashDistribute operator.

Since, we cannot return Best Expr for the STRICT HASH, so that, the
optimization failed with OptimizeChildren failed, and hence aborted the
optimization.

NEXT: explore why CPhysicalMotionHashDistribute cannot match
StrictHashed, and see if we can find the decision point and let it go
through.

-> WIP: Enable StrictHash distribution to generate parallel append plan

- Enable casting from subclass StrictHash to Hash DistributionSpec
- Enable matching of StrictHash with itself
- Fix the distribution requirement order to request StrictHash
  distribution first.

-> Make all children of Append operator rewindable