[#121261025] Derive propagated constraints as Array Expressions
Previously, ORCA propagated array IN constraints (x in (1,2,3)) by first
expanding them into a disjunction of equality statements (x = 1 OR x = 2
...). This process was slow, especially with large IN lists. This change
propagates the array expression. As in, now we derive an array expression,
instead of the more heavyweight OR statement.

A new traceflag has been introduced to enable this feature as in its current
state, it will break several tests involving partitioning.

Some of the unit tests are enabling the array propagation using a trace
flag. This is because some tests are related to a previous story (array expansion
GUCs) and some are related to partitioned tables.