[#121261025] Derive propagated constraints as Array Expressions
Previously, ORCA propagated array IN constraints (x in (1,2,3)) by first
expanding them into a disjunction of equality statements (x = 1 OR x = 2
...). This process was slow, especially with large IN lists. This change
propagates the array expression. As in, now we derive an array expression,
instead of the more heavyweight OR statement.

A new traceflag has been introduced to enable this feature as in its current
state, it will break several tests involving partitioning. A separate
commit will hold the tests.