[#123259241] Fixes partition filters with array constraints
Partition filters are expressions which limit the partitions which GPDB
scans when doing a table scan.

Allowing array constraints to be derived on expressions exposed a situation
where array constraints propagated to partition filters. Arrays weren't
previously handled by the Expression to DXL translator, they were processed
into disjunctions just prior to translation.. The fix we implemented
is twofold:

(1) Allow array expressions to be expanded out into OR expressions during DXL
translation. This avoids any error conditions when array expressions appear
in partition filters because the DXL translator can successfully translate
it.
(2) Do not expand array expressions in a PredicateUtils helper function which
was ran as part of an expression tree pre-DXL translation cleaning.

A traceflag has been added to enable the PredicateUtils change. A test has
been modified to show the effect of enabling the traceflag.

When the traceflag is eventually removed, there will be several more tests
that need updating. This will come in a later PR.